\name{computeValue}
\alias{computeValue}
\title{Internal Function for computeValue}
\description{internal function not called by user}
\usage{
computeValue(Lst,
             delta)
}
\arguments{
  \item{Lst}{list of groups or experiments}
  \item{delta}{an option for changing the minimizing range for the EL
    ratio test statistic}
}
\details{
  This function is used in the dbELnorm function to get the test
  statistic. 
}
\value{
  returns test statistic
}
\note{internal function not called by user}
\author{Lori A. Shepherd, Wan-Min Tsai, Albert Vexler, Jeffrey C. Miecznikowski}
\keyword{internal}
