\name{dbEmpLikeNorm-package}
\docType{package}
\alias{dbEmpLikeNorm-package}
\alias{EmpLikeNorm}
\title{Empirical Likelihood Joint Assessment of Normality}
\description{
Package that has functions to perform a joint assessment of normality
across \emph{k} groups
}
\references{
  Vexler A, Gurevich G, Empirical likelihood ratios applied to
  goodness-of-fit tests based on sample entropy.  Computational
  Statistics and Data Analysis 54(2010) 531-545.

  Gurevich G, Vexler A, A two-sample empirical likelihood ratio test
  based on samples entropy.  Statistics and Computing, 2011.

  Miecznikowski J,Vexler A, Shepherd L, dbEmpLikeGOF: An R Package for
  Nonparametric Likelihood Ratio Tests for Goodness-of-Fit and Two
  Sample Comparisons Based on Sample Entropy.Journal of Statistical
  Software 2013 (Accepted; to appear)

   Tsai WM, Shepherd LA, Miecznikowski J, Hutson A, Vexler A. (2013). 
   An EL based test for normality in multiple groups. Department of 
   Biostatistics. University at Buffalo. Report 1204.

}
\seealso{\code{\link{dbELnorm}}, \code{\link{returnCutoffValue}}}
\author{Lori A. Shepherd, Wan-Min Tsai, Albert Vexler, Jeffrey C. Miecznikowski}
\keyword{package}
